Reorganizing monitord structure with triggers logic
